Moved the helper to interfaces/builtin/
