+++ interfaces/builtin/pulseaudio_test.go
@@ -107,5 +107,5 @@ func (s *PulseAudioInterfaceSuite) TestSecCompOnAllSnaps(c *C) {
 	c.Assert(err, IsNil)
 	err = seccompSpec.AddConnectedPlug(s.iface, s.plug, s.slot)
 	c.Assert(err, IsNil)
-	testutil.NewSecCompSpecChecker(c, seccompSpec, 2).Contains("snap.pulseaudio.app1", "listen\n").Contains("snap.other.app2", "shmctl\n")
+	builtin.NewSecCompSpecChecker(c, seccompSpec, 2).Contains("snap.pulseaudio.app1", "listen\n").Contains("snap.other.app2", "shmctl\n")
 }
