--- testutil/ifacespec_checks.go
@@ -1,53 +1,0 @@
-// -*- Mode: Go; indent-tabs-mode: t -*-
-
-/*
- * Copyright (C) 2017 Canonical Ltd
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License version 3 as
- * published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package testutil
-
-import (
-	. "gopkg.in/check.v1"
-
-	"github.com/snapcore/snapd/interfaces/seccomp"
-	"strings"
-)
-
-type SecCompSpecChecker struct {
-	c        *C
-	snippets map[string]string
-}
-
-func NewSecCompSpecChecker(c *C, spec *seccomp.Specification, expectedLen int) *SecCompSpecChecker {
-	origSnippets := spec.Snippets()
-	c.Assert(len(origSnippets), Equals, expectedLen)
-
-	// flatten the snippets per-security tag for easy testing
-	snippets := make(map[string]string)
-	for k, v := range origSnippets {
-		snippets[k] = strings.Join(v, "\n")
-	}
-	return &SecCompSpecChecker{
-		c:        c,
-		snippets: snippets,
-	}
-}
-
-func (s *SecCompSpecChecker) Contains(securityTag string, val string) *SecCompSpecChecker {
-	s.c.Assert(s.snippets[securityTag], NotNil)
-	s.c.Check(s.snippets[securityTag], Contains, val)
-	return s
-}
